requirements.txt
Flask
gunicorn
sklearn
scikit-learn
pandas
ipykernel
PyYAML
dill
-e .
